+++ SpringBootExample/build-sampleapp-image.sh
@@ -11,7 +11,7 @@ docker run \
     --volume "${PWD}/maven_cache":/root/.m2 \
     --workdir "${PWD}" \
     tools \
-        mvn install:install-file \
+        mvn --batch-mode install:install-file \
         -Dfile=spring-boot-conjur.jar \
         -DgroupId=com.cyberark.conjur.springboot \
         -DartifactId=Spring-boot-conjur \
@@ -25,14 +25,14 @@ docker run \
     --volume "${PWD}/maven_cache":/root/.m2 \
     --workdir "${PWD}" \
     tools \
-        mvn -f pom.xml versions:use-dep-version -Dincludes=com.cyberark.conjur.springboot:Spring-boot-conjur -DdepVersion="$(<../VERSION)"
+        mvn --batch-mode -f pom.xml versions:use-dep-version -Dincludes=com.cyberark.conjur.springboot:Spring-boot-conjur -DdepVersion="$(<../VERSION)"
 
 docker run \
     --volume "${PWD}:${PWD}" \
     --volume "${PWD}/maven_cache":/root/.m2 \
     --workdir "${PWD}" \
     tools \
-        mvn -f pom.xml dependency:tree
+        mvn --batch-mode -f pom.xml dependency:tree
 
 # Use Tools image to package code
 docker run \
@@ -40,7 +40,7 @@ docker run \
     --volume "${PWD}/maven_cache":/root/.m2 \
     --workdir "${PWD}" \
     tools \
-        mvn -f pom.xml clean compile package
+        mvn --batch-mode -f pom.xml clean compile package
 
 docker build \
     --no-cache \