+++ CHANGELOG.md
@@ -6,8 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Nothing should go in this section, please add to the latest unreleased version
+  (and update the corresponding date), or add a new version.
 
-## [0.0.1] 2022-02-03
+## [1.0.0] - 2022-04-04
+
+### Added
+- Conjur Enterprise Tests
+- Publication to Maven Central
+- Code Signing
+
+## [0.0.1] - 2022-02-03
 
 ### Added
 - Initial Code Import